requirements.txt
streamlit
numpy
pillow
opencv-python-headless
gdown
tensorflow==2.15.0